requirements/ci.txt
-r wheel.txt
pip==9.0.1
flake8==3.4.1
pyflakes>=1.0.0
coverage==4.4.1
sphinx==1.6.4
alabaster>=0.6.2
pytest-cov==2.5.1
pygments==2.2.0
mypy==0.521; (python_version>="3.5" or platform_system != "Windows") and platform_python_implementation != "PyPy"
git+git://github.com/python/mypy.git@b6bcdc0; python_version<"3.5" and platform_system == "Windows"
-e .

codecov==2.0.9; platform_system!="Windows"  # We only use it in Travis CI
